+++ misc_testsuite/empty.wast
@@ -1,0 +1,3 @@
+(module (func (export "empty")))
+
+(invoke "empty")
